requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.0
requests==2.32.3
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.10
pydantic==2.10.3
